move logger test
