+++ test/unit/logger/logger.test.js
@@ -1,12 +1,12 @@
 import ConsoleTransportInstance from 'winston-transport'
-import { AzureEventHubTransport } from '../../app/logger/AzureEventHubTransport.js'
+import { AzureEventHubTransport } from '../../../app/logger/AzureEventHubTransport.js'
 
 describe('logger', () => {
   it('Single default log transport enabled', async () => {
     process.env.EVENT_HUB_DISABLED = 'true'
     process.env.APPINSIGHTS_CONNECTIONSTRING = undefined
     const { logger } = await import(
-      `../../app/logger/logger.js?version=${Date.now()}`
+      `../../../app/logger/logger.js?version=${Date.now()}`
     )
     expect(logger.transports.length).toEqual(1)
     expect(logger.transports[0]).toBeInstanceOf(ConsoleTransportInstance)
@@ -15,7 +15,7 @@ describe('logger', () => {
   it('Two log transport enabled including AzureEventHubTransport', async () => {
     process.env.EVENT_HUB_DISABLED = 'false'
     const { logger } = await import(
-      `../../app/logger/logger.js?version=${Date.now()}`
+      `../../../app/logger/logger.js?version=${Date.now()}`
     )
     expect(logger.transports.length).toEqual(2)
     expect(logger.transports[0]).toBeInstanceOf(ConsoleTransportInstance)
